# tests/hospital-api-v1_2_openapi.yaml
openapi: 3.1.0
info:
  title: Hospital Management API
  version: 1.2.0
  description: |
    Patients, doctors, departments, OPD schedules, appointments, and the
    hospital information the call centre's voice agent reads to callers.

    This API is the SINGLE SOURCE for everything the voice agent says.
    Structured data (schedules, availability, prices, hours, rooms) is read
    live at the moment a caller asks. Prose (department descriptions,
    services, FAQ articles) is authored here and copied into the agent's
    search index by a sync job that polls `updated_since` (see "Sync").

    ## What changed in 1.2.0

    - NEW `Holidays` — hospital-wide closures. Availability returns a
      `holiday` status and booking refuses the date with `409 HOLIDAY`.
    - NEW `visit_type` on appointments (`new`, `follow_up`,
      `report_showing`) and an optional separate report-showing queue per
      session (`report_capacity`).
    - NEW `POST /schedules/import` — upload the OPD room-allocation
      spreadsheet (or CSV), preview the result, then apply it.
    - NEW `GET /audit-log` — who changed what, when, with before/after.
    - NEW `Hospital` profile — address, hotlines, emails, OPD and desk
      hours (replaces the voice agent's `CRM_OPENING_HOURS` setting).
    - NEW `Packages`, `Services` (including facility rooms) and `Articles`
      (FAQ and general content).
    - CHANGED `Department` — adds `description_bn` and `conditions`.
...

    ## What changed in 1.1.0

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - APIKeyHeader: []
...
paths:
...
  /departments:
    get:
      tags: [Departments]
      summary: List Departments
      operationId: list_departments_departments_get
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          required: false
          schema: { type: [string, 'null'] }
          description: |
            Case-insensitive match against `name`, `name_bn` and every entry
            of `aliases`, by word prefix (so `ent` matches "ENT, Head & Neck
            Surgery" but not "Gastroenterology").
        - name: service_line
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/ServiceLine'
              - type: 'null'
        - name: is_active
          in: query
          required: false
          schema: { type: [boolean, 'null'], default: true }
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/IncludeDeleted'
      responses:
        '200':
          description: Successful Response
          headers:
            X-Server-Time: { $ref: '#/components/headers/XServerTime' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DepartmentOut' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Departments]
      summary: Create Department
      operationId: create_department_departments_post
      description: Requires `admin` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: '`DUPLICATE_NAME`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /departments/{department_id}:
    parameters:
      - name: department_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Departments]
      summary: Get Department
      operationId: get_department_departments__department_id__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentOut' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Departments]
      summary: Update Department
      operationId: update_department_departments__department_id__put
      description: Requires `admin` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentUpdate' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Departments]
      summary: Delete Department
      operationId: delete_department_departments__department_id__delete
      description: |
        Requires `admin` scope. Refused with `409 HAS_DOCTORS` while any doctor
        references it — set `is_active: false` instead. CHANGED in 1.2.0:
        soft delete (sets `deleted_at`).
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`HAS_DOCTORS`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ── Doctors ──────────────────────────────────────────────────────────────
  /doctors:
    post:
      tags: [Doctors]
      summary: Create Doctor
      operationId: create_doctor_doctors_post
      description: Requires `admin` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DoctorCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DoctorOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [Doctors]
      summary: List Doctors
      operationId: list_doctors_doctors_get
      description: |
        CHANGED in 1.1.0: new filters `department_id`, `is_active`,
        `available_on`; `search` also matches `name_bn`.
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: specialization
          in: query
          required: false
          schema: { type: [string, 'null'] }
          description: Substring match on the display `specialization`. Prefer `department_id`.
        - name: search
          in: query
          required: false
          schema: { type: [string, 'null'] }
          description: Substring match on `name` and `name_bn`. `%` and `_` are matched literally.
        - name: department_id
          in: query
          required: false
          schema: { type: [integer, 'null'] }
          description: NEW in 1.1.0.
        - name: is_active
          in: query
          required: false
          schema: { type: [boolean, 'null'], default: true }
          description: NEW in 1.1.0. Pass `null`/omit explicitly to include inactive doctors (admin views).
        - name: available_on
          in: query
          required: false
          schema: { type: [string, 'null'], format: date }
          description: |
            NEW in 1.1.0. Only doctors with at least one session on that date
            after applying exceptions and holidays (status `available` or
            `full` in availability terms).
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/IncludeDeleted'
      responses:
        '200':
          description: Successful Response
          headers:
            X-Server-Time: { $ref: '#/components/headers/XServerTime' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DoctorOut' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /doctors/{doctor_id}:
    parameters:
      - $ref: '#/components/parameters/DoctorId'
    get:
      tags: [Doctors]
      summary: Get Doctor
      operationId: get_doctor_doctors__doctor_id__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DoctorOut' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    put:
      tags: [Doctors]
      summary: Update Doctor
      operationId: update_doctor_doctors__doctor_id__put
      description: Requires `admin` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DoctorUpdate' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DoctorOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Doctors]
      summary: Delete Doctor
      operationId: delete_doctor_doctors__doctor_id__delete
      description: |
        Requires `admin` scope. CHANGED in 1.1.0: refused with
        `409 HAS_FUTURE_APPOINTMENTS` while the doctor has scheduled
        appointments from today onward — set `is_active: false` instead.
        CHANGED in 1.2.0: soft delete (sets `deleted_at`), so past
        appointments keep their doctor.
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`HAS_FUTURE_APPOINTMENTS`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ── Schedules (NEW) ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Skip:
      name: skip
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
    DoctorId:
      name: doctor_id
      in: path
      required: true
      schema: { type: integer }
...
    UpdatedSince:
      name: updated_since
      in: query
      required: false
      schema: { type: [string, 'null'], format: date-time }
      description: |
        NEW in 1.2.0. Only records created, updated or deleted after this
        instant. Pass the previous response's `X-Server-Time`.
    IncludeDeleted:
      name: include_deleted
      in: query
      required: false
      schema: { type: boolean, default: false }
      description: |
        NEW in 1.2.0. Also return soft-deleted, inactive and unpublished
        records (each with `deleted_at` / `is_active` / `is_published` set),
        so a sync job can remove them from its index.
  headers:
    XServerTime:
      description: NEW in 1.2.0. Server time when the list was read; the next `updated_since`.
      schema: { type: string, format: date-time }
  responses:
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: '`FORBIDDEN_SCOPE` — the API key''s scope does not allow this operation.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      description: NEW in 1.1.0. Body of every business-rule error.
      required: [code, detail]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: Human-readable, English. Not for parsing.
        next_available_date:
          type: [string, 'null']
          format: date
          description: |
            On `NO_SESSION`, `SESSION_FULL` and `HOLIDAY`: the doctor's next
            date (within 31 days) with a free serial, so a client can offer
            it directly.
        holiday_name:
          type: [string, 'null']
          description: NEW in 1.2.0. On `HOLIDAY`.
        candidates:
          type: [array, 'null']
          items: { type: object }
          description: Optional supporting rows (e.g. the clashing schedule on `SCHEDULE_OVERLAP`).
    ErrorCode:
      type: string
      enum:
        - NOT_FOUND
        - PATIENT_NOT_FOUND
        - DOCTOR_NOT_FOUND
        - FORBIDDEN_SCOPE
        - PAST_DATE
        - NO_SESSION
        - SESSION_REQUIRED
        - SESSION_FULL
        - DUPLICATE_BOOKING
        - DOCTOR_NOT_ACCEPTING
        - SCHEDULE_OVERLAP
        - ROOM_CONFLICT
        - HAS_DOCTORS
        - HAS_FUTURE_APPOINTMENTS
        - DUPLICATE_NAME
        - IDEMPOTENCY_KEY_REUSED
        - HOLIDAY
        - DUPLICATE_DATE
        - IMPORT_HAS_ERRORS
        - UNSUPPORTED_FILE
...
    ServiceLine:
      type: string
      description: Which call-centre desk handles this department's calls.
      enum: [appointment, emergency, support]
...
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf: [{ type: string }, { type: integer }]
        msg: { type: string }
        type: { type: string }

    # ── patients (unchanged shape) ─────────────────────────────────────────
...
    DepartmentBase:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
          description: English display name, e.g. "Obstetrics & Gynaecology".
        name_bn:
          type: [string, 'null']
          maxLength: 150
          description: Bangla name, as spoken to callers, e.g. "অবস্টেট্রিক্স ও গাইনোকোলজি".
        aliases:
          type: array
          items: { type: string, maxLength: 80 }
          default: []
          description: |
            Other words callers use, in either language and any spelling:
            ["gynae", "gynecology", "gynaecologist", "স্ত্রীরোগ"]. Matched by
            `GET /departments?search=`.
        service_line: { $ref: '#/components/schemas/ServiceLine' }
        description:
          type: [string, 'null']
          description: |
            What the department does, in English, in a few short paragraphs
            a caller can be read. No staff contact details. Synced to the
            voice agent's search index.
        description_bn:
          type: [string, 'null']
          description: NEW in 1.2.0. The same, in Bangla.
        conditions:
          type: array
          items: { type: string, maxLength: 120 }
          default: []
          description: |
            NEW in 1.2.0. Complaints and conditions this department sees, in
            the words a caller uses, either language: ["kidney stone",
            "dialysis", "কিডনি", "প্রস্রাবে জ্বালা"]. Lets the agent route
            "which department for my kidney problem?".
        opening_hours:
          type: [string, 'null']
          maxLength: 200
          description: Free text, e.g. "8am to 8pm, Saturday to Thursday" or "24 hours".
        is_active: { type: boolean, default: true }
    DepartmentCreate:
      allOf:
        - $ref: '#/components/schemas/DepartmentBase'
        - type: object
          required: [name, service_line]
    DepartmentUpdate:
      type: object
      description: Every field optional; omitted fields are unchanged.
      properties:
        name: { type: [string, 'null'], maxLength: 150 }
        name_bn: { type: [string, 'null'], maxLength: 150 }
        aliases:
          type: [array, 'null']
          items: { type: string, maxLength: 80 }
        service_line:
          anyOf:
            - $ref: '#/components/schemas/ServiceLine'
            - type: 'null'
        description: { type: [string, 'null'] }
        description_bn: { type: [string, 'null'] }
        conditions:
          type: [array, 'null']
          items: { type: string, maxLength: 120 }
        opening_hours: { type: [string, 'null'], maxLength: 200 }
        is_active: { type: [boolean, 'null'] }
    DepartmentOut:
      allOf:
        - $ref: '#/components/schemas/DepartmentBase'
        - type: object
          required: [id, name, service_line, is_active, created_at]
          properties:
            id: { type: integer }
            doctor_count:
              type: integer
              description: Active doctors in this department.
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
            deleted_at: { type: [string, 'null'], format: date-time, description: NEW in 1.2.0. }

    # ── doctors ────────────────────────────────────────────────────────────
    DoctorBase:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
          description: Full name with honorifics as the hospital writes it.
        name_bn:
          type: [string, 'null']
          maxLength: 150
          description: NEW in 1.1.0. Bangla spelling of the name.
        designation: { type: [string, 'null'], description: 'e.g. "Senior Consultant"' }
        education: { type: [string, 'null'] }
        specialization:
          type: [string, 'null']
          description: Display text only. Use `department_id` for filtering.
        department_id:
          type: [integer, 'null']
          description: NEW in 1.1.0.
        description: { type: [string, 'null'] }
        mobile_number:
          type: [string, 'null']
          description: Staff-private. Omitted from responses for `agent`-scoped keys.
        email:
          type: [string, 'null']
          description: Staff-private. Omitted from responses for `agent`-scoped keys.
        is_active:
          type: boolean
          default: true
          description: NEW in 1.1.0. False for doctors who have left; excluded from lists by default.
        accepting_appointments:
          type: boolean
          default: true
          description: |
            NEW in 1.1.0. False blocks new bookings (`409
            DOCTOR_NOT_ACCEPTING`) while keeping the doctor listed.
        consultation_fee:
          type: [number, 'null']
          minimum: 0
          description: NEW in 1.1.0. In BDT. Null means not published.
    DoctorCreate:
      allOf:
        - $ref: '#/components/schemas/DoctorBase'
        - type: object
          required: [name]
    DoctorUpdate:
      type: object
      description: Every field optional; omitted fields are unchanged.
      properties:
        name: { type: [string, 'null'], maxLength: 150 }
        name_bn: { type: [string, 'null'], maxLength: 150 }
        designation: { type: [string, 'null'] }
        education: { type: [string, 'null'] }
        specialization: { type: [string, 'null'] }
        department_id: { type: [integer, 'null'] }
        description: { type: [string, 'null'] }
        mobile_number: { type: [string, 'null'] }
        email: { type: [string, 'null'] }
        is_active: { type: [boolean, 'null'] }
        accepting_appointments: { type: [boolean, 'null'] }
        consultation_fee: { type: [number, 'null'], minimum: 0 }
    DoctorOut:
      allOf:
        - $ref: '#/components/schemas/DoctorBase'
        - type: object
          required: [id, name, is_active, accepting_appointments, created_at]
          properties:
            id: { type: integer }
            department:
              anyOf:
                - $ref: '#/components/schemas/DepartmentSummary'
                - type: 'null'
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
            deleted_at: { type: [string, 'null'], format: date-time, description: NEW in 1.2.0. }
...
    DepartmentSummary:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        name_bn: { type: [string, 'null'] }

    # ── schedules (NEW) ────────────────────────────────────────────────────
...